refactor the projects with dapper

+++ Arsenalcn.Core/Extension/RepositoryExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Data.SqlClient;
-using System.Diagnostics.Contracts;
 using System.Linq;
 
 namespace Arsenalcn.Core
@@ -9,15 +8,16 @@ namespace Arsenalcn.Core
     {
         public static int Insert<T>(this IEnumerable<T> source, SqlTransaction trans = null) where T : class, IEntity
         {
-            Contract.Requires(source != null);
-
-            IRepository repo = new Repository();
-
             var list = source as IList<T> ?? source.ToList();
 
-            foreach (var instance in list)
+            if (list.Count > 0)
             {
-                repo.Insert((IDao)instance, trans);
+                IRepository repo = new Repository();
+
+                foreach (var instance in list)
+                {
+                    repo.Insert((IDao)instance, trans);
+                }
             }
 
             return list.Count;
@@ -25,15 +25,16 @@ namespace Arsenalcn.Core
 
         public static int Update<T>(this IEnumerable<T> source, SqlTransaction trans = null) where T : class, IEntity
         {
-            Contract.Requires(source != null);
-
-            IRepository repo = new Repository();
-
             var list = source as IList<T> ?? source.ToList();
 
-            foreach (var instance in list)
+            if (list.Count > 0)
             {
-                repo.Update(instance, trans);
+                IRepository repo = new Repository();
+
+                foreach (var instance in list)
+                {
+                    repo.Update(instance, trans);
+                }
             }
 
             return list.Count;
@@ -41,15 +42,16 @@ namespace Arsenalcn.Core
 
         public static int Delete<T>(this IEnumerable<T> source, SqlTransaction trans = null) where T : class, IEntity
         {
-            Contract.Requires(source != null);
-
-            IRepository repo = new Repository();
-
             var list = source as IList<T> ?? source.ToList();
 
-            foreach (var instance in list)
+            if (list.Count > 0)
             {
-                repo.Delete(instance, trans);
+                IRepository repo = new Repository();
+
+                foreach (var instance in list)
+                {
+                    repo.Delete(instance, trans);
+                }
             }
 
             return list.Count;
